Refactor course list and empty state (#308)
