+++ package.json
@@ -33,11 +33,11 @@
     "@edx/frontend-platform": "7.1.0",
     "@edx/openedx-atlas": "^0.6.0",
     "@edx/react-unit-test-utils": "2.0.0",
-    "@openedx/frontend-plugin-framework": "1.0.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-brands-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.15",
+    "@openedx/frontend-plugin-framework": "^1.0.2",
     "@openedx/paragon": "^21.11.3",
     "@optimizely/react-sdk": "^2.9.2",
     "@redux-beacon/segment": "^1.1.0",
