+++ src/containers/CoursesPanel/CourseList/hooks.js
@@ -1,0 +1,8 @@
+import { useWindowSize, breakpoints } from '@openedx/paragon';
+
+export const useIsCollapsed = () => {
+  const { width } = useWindowSize();
+  return width < breakpoints.medium.maxWidth;
+};
+
+export default useIsCollapsed;
