+++ src/containers/CoursesPanel/NoCoursesView/index.jsx
