+++ src/containers/CoursesPanel/NoCoursesView/index.scss
