+++ src/containers/CoursesPanel/NoCoursesView/index.test.jsx
