+++ src/containers/CoursesPanel/NoCoursesView/messages.js
