+++ src/containers/CoursesPanel/hooks.test.js
