+++ src/containers/CoursesPanel/index.scss
