+++ src/containers/CoursesPanel/messages.js
