+++ src/containers/Dashboard/index.jsx
@@ -4,7 +4,7 @@ import { reduxHooks } from 'hooks';
 import { RequestKeys } from 'data/constants/requests';
 import EnterpriseDashboardModal from 'containers/EnterpriseDashboardModal';
 import SelectSessionModal from 'containers/SelectSessionModal';
-import CourseList from 'containers/CourseList';
+import CoursesPanel from 'containers/CoursesPanel';
 
 import LoadedSidebar from 'containers/WidgetContainers/LoadedSidebar';
 import NoCoursesSidebar from 'containers/WidgetContainers/NoCoursesSidebar';
@@ -36,7 +36,7 @@ export const Dashboard = () => {
           ? (<LoadingView />)
           : (
             <DashboardLayout sidebar={hasCourses ? LoadedSidebar : NoCoursesSidebar}>
-              <CourseList />
+              <CoursesPanel />
             </DashboardLayout>
           )}
       </div>
