+++ src/containers/Dashboard/index.test.jsx
@@ -4,7 +4,7 @@ import { reduxHooks } from 'hooks';
 
 import EnterpriseDashboardModal from 'containers/EnterpriseDashboardModal';
 import SelectSessionModal from 'containers/SelectSessionModal';
-import CourseList from 'containers/CourseList';
+import CoursesPanel from 'containers/CoursesPanel';
 
 import LoadedWidgetSidebar from 'containers/WidgetContainers/LoadedSidebar';
 import NoCoursesWidgetSidebar from 'containers/WidgetContainers/NoCoursesSidebar';
@@ -24,7 +24,7 @@ jest.mock('hooks', () => ({
 }));
 
 jest.mock('containers/EnterpriseDashboardModal', () => 'EnterpriseDashboardModal');
-jest.mock('containers/CourseList', () => 'CourseList');
+jest.mock('containers/CoursesPanel', () => 'CoursesPanel');
 jest.mock('containers/WidgetContainers/LoadedSidebar', () => 'LoadedWidgetSidebar');
 jest.mock('containers/WidgetContainers/NoCoursesSidebar', () => 'NoCoursesWidgetSidebar');
 jest.mock('./LoadingView', () => 'LoadingView');
@@ -116,7 +116,7 @@ describe('Dashboard', () => {
           showSelectSessionModal: true,
         },
         content: ['LoadedView', (
-          <DashboardLayout sidebar={LoadedWidgetSidebar}><CourseList /></DashboardLayout>
+          <DashboardLayout sidebar={LoadedWidgetSidebar}><CoursesPanel /></DashboardLayout>
         )],
         showEnterpriseModal: false,
         showSelectSessionModal: true,
@@ -132,7 +132,7 @@ describe('Dashboard', () => {
           showSelectSessionModal: false,
         },
         content: ['Dashboard layout with no courses sidebar and content', (
-          <DashboardLayout sidebar={NoCoursesWidgetSidebar}><CourseList /></DashboardLayout>
+          <DashboardLayout sidebar={NoCoursesWidgetSidebar}><CoursesPanel /></DashboardLayout>
         )],
         showEnterpriseModal: true,
         showSelectSessionModal: false,
